+++ __init__.py
@@ -15,3 +15,4 @@ Migrate(app, db)
 # Setup LoginManager object (app)
 login_manager = LoginManager()
 login_manager.init_app(app)
+next_page = None
