+++ cruddy/templates/cruddy/login.html
@@ -1,30 +1,33 @@
 <!DOCTYPE html>
+
+
+
 {% block content %}
 
-    <div class="container py-4">
-        <header class="pb-3 mb-4 border-bottom border-primary text-dark">
-            <span class="fs-4">SQL CRUD Login Page </span>
-        </header>
-    </div>
+<div class="container py-4">
+    <header class="pb-3 mb-4 border-bottom border-primary text-dark">
+        <span class="fs-4">SQL CRUD Login Page</span>
+    </header>
+</div>
 
-    <div class="container py-4 text-light bg-success">
+<div class="container py-4 text-light bg-success">
 
-        <div class="container bg-secondary py-4">
-            <div class="p-5 mb-4 bg-light text-dark rounded-3">
-                <form method="POST" ID="login" action={{url_for('crud.crud_login')}}> <!--originally crud.crud_login-->
-                    <table>
-                        <tr><th><label for="email">Email</label></th></tr>
-                        <tr><td><input type="email" id='email' name="email" size="20" required></td></tr>
-                        <tr><th><label for="password">Password</label></th></tr>
-                        <tr><td><input type="password" id="password" name="password" size="20" required></td></tr>
-                        <tr><th><input type="submit" value="Submit"></th></tr>
-                        <tr><td><a href={{url_for('crud.crud_authorize')}}>Sign Up</a></td></tr> <!--originally crud_async-->
-                    </table>
-                </form>
-            </div>
+    <div class="container bg-secondary py-4">
+        <div class="p-5 mb-4 bg-light text-dark rounded-3">
+            <form method="POST" ID="login" action={{url_for('main_login')}}>
+                <table>
+                    <tr><th><label for="email">Email</label></th></tr>
+                    <tr><td><input type="email" id='email' name="email" size="20" required></td></tr>
+                    <tr><th><label for="password">Password</label></th></tr>
+                    <tr><td><input type="password" id="password" name="password" size="20" required></td></tr>
+                    <tr><th><input type="submit" value="Submit"></th></tr>
+                    <tr><td><a href={{url_for('main_authorize')}}>Sign Up</a></td></tr>
+                </table>
+            </form>
         </div>
-
     </div>
+
+</div>
 
 
 {% endblock %}
