+++ main.py
@@ -1,11 +1,14 @@
-from flask import Flask, render_template, request
-from __init__ import app
-import requests
+from flask import Flask, render_template, request, redirect, url_for
+from __init__ import app, login_manager
+from flask_login import login_required
+
 
 from cruddy.app_crud import app_crud
 
 
 from pathlib import Path
+
+from cruddy.login import logout
 
 app.register_blueprint(app_crud)
 
@@ -16,6 +19,61 @@ def index():
     return render_template("index.html")
 
 
+# Flask-Login directs unauthorised users to this unauthorized_handler
+@login_manager.unauthorized_handler
+def unauthorized():
+    """Redirect unauthorized users to Login page."""
+    global next_page
+    next_page = request.endpoint
+    return redirect(url_for('main_login'))
+
+
+# if login url, show phones table only
+@app.route('/login/', methods=["GET", "POST"])
+def main_login():
+    # obtains form inputs and fulfills login requirements
+    global next_page
+    if request.form:
+        email = request.form.get("email")
+        password = request.form.get("password")
+        if login(email, password):
+            try:        # try to redirect to next page
+                temp = next_page
+                next_page = None
+                return redirect(url_for(temp))
+            except:     # any failure goes to home page
+                return redirect(url_for('index'))
+
+
+    # if not logged in, show the login page
+    return render_template("login.html")
+
+
+# if login url, show phones table only
+@app.route('/logout/', methods=["GET", "POST"])
+@login_required
+def main_logout():
+    logout()
+    return redirect(url_for('index'))
+
+
+@app.route('/authorize/', methods=["GET", "POST"])
+def main_authorize():
+    error_msg = ""
+    # check form inputs and creates user
+    if request.form:
+        # validation should be in HTML
+        user_name = request.form.get("user_name")
+        email = request.form.get("email")
+        password1 = request.form.get("password1")
+        password2 = request.form.get("password2")  # password should be verified
+        if password1 == password2:
+            if authorize(user_name, email, password1):
+                return redirect(url_for('main_login'))
+        else:
+            error_msg = "Passwords do not match"
+    # show the auth user page if the above fails for some reason
+    return render_template("authorize.html", error_msg=error_msg)
 # connects /kangaroos path to render kangaroos.html
 @app.route('/kangaroos/')
 def kangaroos():
@@ -113,7 +171,7 @@ def bookapi():
 
 # runs the application on the development server
 if __name__ == "__main__":
-    app.run(debug=True,port=9000) #says "run this directly" app.run will run the server
+    app.run(debug=True,port=8000) #says "run this directly" app.run will run the server
 
 #index.html is standard
